app/requirements.txt
appdirs==1.4.3
attrs==20.3.0
bcrypt==3.2.0
CacheControl==0.12.6
cachelib==0.1.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Session==0.3.2
html5lib==1.0.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient==2.0.2
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
protobuf==3.14.0
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.6
pytest==6.2.1
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.22
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
